requirements.txt
websockets
openai
requests
Pillow
wordcloud
matplotlib
pyaudio
assemblyai
PyYAML
pillow
keyboard
